feat: add find(full) function
Well, I don't know why I call this function full but there's a story I
must tell regarding this function.

This is the most optimal result that I come up with. The reason is, my
previous attempts created monstrous outputs such as
```rust
[
	{
		MangaName: "something",
		... rest of the manga info
	},
	{
		StaffName: "something",
		... rest of the staff info,
	},
	{
		StaffName: "Another staff",
	},
]
```
and so on. As you can see, this format is horrendous. Now with the
`Resp` struct it looks more pretty and it become parsable as
well. Until I find a better way to handle this, this implementation
will be the solution.
Signed-off-by: Taylan Dogan <doganntaylan@gmail.com>